+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Enhancement: The step size is now synchronized between the main screen and the Probing screen
 - Enhancement: Add Auto Blow, Auto Bed Clean, and Ionizer toggles to the Config and Run screen on Z1
 - Enhancement: Add Z1 bed background
+- Enhancement: Machine bed background images in the config-n-run preview screen now filtered to show images that match the machine model connected
 - Change: Hide Auto Vacuum on the Config and Run screen when the machine is not a C1
 - Change: Config and Run preview now uses now uses the configured worksize_x/y for the bed size
 - Fixed: Y+/Y- jogging buttons on the Probing screen respect the configured Y axis inversion setting